Add new fingerprint module
Doesn't really do anything, but shows that the driver and display module can coexist.
