+++ src/modules/CMakeLists.txt
@@ -7,3 +7,4 @@
 target_include_directories(app PRIVATE .)
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/modules_common.c)
 target_sources_ifdef(CONFIG_DISPLAY_MODULE app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/display_module.c)
+target_sources(app PRIVATE fingerprint_module.c)
